Move vp8 prediction related functions to their own file
