+++ src/lib.rs
@@ -25,6 +25,7 @@ mod lossless_transform;
 mod transform;
 mod vp8_arithmetic_decoder;
 mod vp8_common;
+mod vp8_prediction;
 mod yuv;
 
 pub mod vp8;
